feat: 微信小程序 + 前端MQTT修复 + 暖色系UI美化
## 新增
- 微信小程序 (48文件) — 仪表盘/设备监控/灌溉控制/告警中心
- 后端微信登录接口 POST /auth/wechat-login
- 前端 Buffer polyfill 修复 MQTT WebSocket 连接

## 改动
- 后端: wechat_openid 列, 华为云IoT集成, 串口网关, 告警/策略引擎
- 前端: MQTT store 自动重连, Dashboard并行请求, 实时数据更新
- 设备: 固件适配华为云, config_store Flash存储, SetConfig命令

## 美化
- 小程序暖色系(琥珀金 #f59e0b) 全面替代冷色
- 点击缩放动效、卡片阴影层次、渐变发光条
- 脉冲/淡入/声波动画

+++ device/src/drivers/relay.h
@@ -3,9 +3,8 @@
 #ifndef RELAY_H
 #define RELAY_H
 
-#include <stdint.h>
 
-void relay_init(uint8_t pin);
+void relay_init(unsigned int pin);
 void relay_on(void);
 void relay_off(void);
 int  relay_state(void);  /* 返回 1:开 0:关 */
